+++ src/main/java/org/apache/pdfbox/pdmodel/font/PDType1Font.java
@@ -276,6 +276,6 @@ public class PDType1Font extends PDSimpleFont
         g2d.setRenderingHint( RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON );
         g2d.setFont( awtFont.deriveFont( at ).deriveFont( fontSize ) );
 
-        g2d.drawString( string, (int)x, (int)y );
+        g2d.drawString( string, x, y );
     }
 }
